@@ -1,8 +1,6 @@
-namespace EPS.Extensions.SiteMapIndex
-{
-    public class Constants
-    {
-        public const string URLSET = "http://www.sitemaps.org/schemas/sitemap/0.9";
+namespace EPS.Extensions.SiteMapIndex;
 
-    }
+public static class Constants
+{
+    public const string URLSET = "http://www.sitemaps.org/schemas/sitemap/0.9";
 }
