@@ -1,12 +1,10 @@
 using System;
 
-namespace EPS.Extensions.SiteMapIndex
+namespace EPS.Extensions.SiteMapIndex;
+
+public class SiteMapException : Exception
 {
-    public class SiteMapException: ApplicationException
-    {
-        public SiteMapException(string message) : base(message){}
-        public SiteMapException(string message, Exception innerException): base(message,innerException){}
-        public SiteMapException()
-        {}
-    }
+    public SiteMapException(string message) : base(message) { }
+    public SiteMapException(string message, Exception innerException) : base(message, innerException) { }
+    public SiteMapException() { }
 }
